requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4
    # via python-escpos
argcomplete==2.0.0
    # via python-escpos
argparse==1.4.0
    # via python-escpos
brotli==1.0.9
    # via fonttools
cachetools==5.2.0
    # via premailer
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via weasyprint
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
cssselect==1.1.0
    # via premailer
cssselect2==0.6.0
    # via weasyprint
cssutils==2.5.1
    # via premailer
flask==2.2.1
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==3.0.10
    # via -r requirements.in
fonttools[woff]==4.34.4
    # via weasyprint
gunicorn==20.1.0
    # via -r requirements.in
html5lib==1.1
    # via weasyprint
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via
    #   libusb
    #   pkg-about
importlib-resources==5.9.0
    # via
    #   libusb
    #   pkg-about
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
libusb==1.0.26b3
    # via -r requirements.in
lxml==4.9.1
    # via premailer
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
numpy==1.23.1
    # via opencv-python-headless
opencv-python-headless==4.6.0.66
    # via -r requirements.in
packaging==21.3
    # via
    #   libusb
    #   pkg-about
pillow==9.2.0
    # via
    #   -r requirements.in
    #   python-escpos
    #   weasyprint
pkg-about==1.0.6
    # via libusb
premailer==3.10.0
    # via yagmail
psycopg2-binary==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydyf==0.2.0
    # via weasyprint
pyparsing==3.0.9
    # via packaging
pyphen==0.12.0
    # via weasyprint
pyserial==3.5
    # via python-escpos
python-barcode==0.14.0
    # via -r requirements.in
python-escpos==2.2.0
    # via -r requirements.in
pyusb==1.2.1
    # via python-escpos
pyyaml==6.0
    # via python-escpos
qrcode==7.3.1
    # via
    #   -r requirements.in
    #   python-escpos
requests==2.28.1
    # via premailer
six==1.16.0
    # via
    #   flask-cors
    #   html5lib
    #   python-escpos
tinycss2==1.1.1
    # via
    #   cssselect2
    #   weasyprint
tomli==2.0.1
    # via pkg-about
urllib3==1.26.11
    # via requests
weasyprint==56.1
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
werkzeug==2.2.1
    # via flask
yagmail==0.15.283
    # via -r requirements.in
zipp==3.8.1
    # via importlib-metadata
zopfli==0.2.1
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# setuptools
